CI: also remove /opt/homebrew
Closes: https://trac.macports.org/ticket/69661

+++ .github/workflows/bootstrap.sh
@@ -61,14 +61,16 @@ endgroup
 
 
 begingroup "Uninstalling Homebrew"
-# Move directories to /opt/off
+# Move directories to /opt/*-off
 echo "Moving directories..."
-sudo mkdir /opt/off
-/usr/bin/sudo /usr/bin/find /usr/local -mindepth 1 -maxdepth 1 -type d -print -exec /bin/mv {} /opt/off/ \;
+sudo mkdir /opt/local-off /opt/homebrew-off
+test ! -d /usr/local || /usr/bin/sudo /usr/bin/find /usr/local -mindepth 1 -maxdepth 1 -type d -print -exec /bin/mv {} /opt/local-off/ \;
+test ! -d /opt/homebrew || /usr/bin/sudo /usr/bin/find /opt/homebrew -mindepth 1 -maxdepth 1 -type d -print -exec /bin/mv {} /opt/homebrew-off/ \;
 
 # Unlink files
 echo "Removing files..."
-/usr/bin/sudo /usr/bin/find /usr/local -mindepth 1 -maxdepth 1 -type f -print -delete
+test ! -d /usr/local || /usr/bin/sudo /usr/bin/find /usr/local -mindepth 1 -maxdepth 1 -type f -print -delete
+test ! -d /opt/homebrew || /usr/bin/sudo /usr/bin/find /opt/homebrew -mindepth 1 -maxdepth 1 -type f -print -delete
 
 # Rehash to forget about the deleted files
 hash -r
